:sparkles: feat: 알림 제목 자동완성 메소드 및 알림 발송 구현

+++ src/main/java/com/plane/comment/dto/CommentDeleteRequest.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 public class CommentDeleteRequest {
 	
 	@NotNull(message = "댓글 번호는 필수입니다.")
-	private Integer commentId;
+	private Long commentId;
 	
 	
 	@NotNull(message = "글 번호는 필수입니다.")
@@ -15,12 +15,12 @@ public class CommentDeleteRequest {
 	public CommentDeleteRequest() {}
 
 
-	public Integer getCommentId() {
+	public Long getCommentId() {
 		return commentId;
 	}
 
 
-	public void setCommentId(Integer commentId) {
+	public void setCommentId(Long commentId) {
 		this.commentId = commentId;
 	}
